camera fixes and improovements
Implemented jitter fix for the camera while running and folowing each object. Moved the camera on the top of the rocket and implemented rotation in order to be able to observe the solar system. Modified the asteroid belt and sphere to be accurate compared to planet position and solar system dimention