final push?
modified earth texture resolution to be more detailed while in the start menu.
modified the audio manager to include a fade out effect when changing planets. modified the planet info script to include the facts as list. modified the select planet script to show each fact on a new line and to play each planet theme after the change of the parent planet is made (before that it was always playing the theme of the last parent planet)